+++ src/components/ui/index.js
@@ -1,4 +1,5 @@
 import BreadcrumbUi from '@/components/ui/BreadcrumbUi'
+import ButtonUi from '@/components/ui/ButtonUi'
 import ConfirmUi from '@/components/ui/ConfirmUi'
 import FormField from '@/components/ui/form/FormField'
 import IconUi from '@/components/ui/IconUi'
@@ -8,6 +9,7 @@ import TabsUi from '@/components/ui/TabsUi'
 
 export {
   BreadcrumbUi,
+  ButtonUi,
   ConfirmUi,
   FormField,
   IconUi,