+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased](../../compare/master...dev)
 ### Added
+- Add active class on admin menu
+- Redirect from admin empty route (/admin/ to /admin/dashboard)
 ### Changed
 ### Deprecated
 ### Removed