+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - Add active class on admin menu
 - Redirect from admin empty route (/admin/ to /admin/dashboard)
+- Add news component FormField
 ### Changed
+- Changed the layoutAdmin grid to prevent jump of content on transition
 ### Deprecated
 ### Removed
 - Removed old props title and subtitle in AdminTable